refactor(novel-card): add ranked layout for NovelCard component with rank badge and styling
