+++ src/app/(public)/top-rated/page.tsx
@@ -1,39 +1,16 @@
 "use client";
 
 import { useState } from "react";
-import {
-  Star,
-  TrendingUp,
-  Clock,
-  BookOpen,
-  Eye,
-  Crown,
-  Trophy,
-  Award,
-  Medal,
-} from "lucide-react";
+import { Star, TrendingUp, Eye, Trophy } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
 import { Skeleton } from "@/components/ui/skeleton";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import Link from "next/link";
-import Image from "next/image";
 
 import { useNovels } from "@/hooks/use-novels";
-import {
-  formatRating,
-  formatChapterCount,
-  formatViewCount,
-  truncateDescription,
-  getStatusColor,
-  getNovelStyling,
-  getNovelBadgeConfig,
-} from "@/lib/novel-utils";
-import { cn } from "@/lib/utils";
 import { Novel } from "@/types/api";
-import { NovelRating } from "@/components/novels";
-import { NovelBadge } from "@/components/novels/ui/novel-badge";
+import { NovelCard } from "@/components/novels";
 
 export default function TopRatedPage() {
   const [activeTab, setActiveTab] = useState("rating");
@@ -52,178 +29,6 @@ export default function TopRatedPage() {
   // Extract data arrays from paginated responses
   const topRated = topRatedData?.data;
   const mostViewed = mostViewedData?.data;
-
-  const getRankIcon = (rank: number) => {
-    if (rank === 1) return <div className="text-xl">1</div>;
-    if (rank === 2) return <div className="text-xl">2</div>;
-    if (rank === 3) return <div className="text-xl">3</div>;
-    return null;
-  };
-
-  const getRankColor = (rank: number) => {
-    if (rank === 1) return "from-yellow-400 via-yellow-500 to-yellow-600";
-    if (rank === 2) return "from-slate-300 via-slate-400 to-slate-500";
-    if (rank === 3) return "from-orange-400 via-orange-500 to-orange-600";
-    if (rank <= 10) return "from-blue-400 via-blue-500 to-blue-600";
-    return "from-purple-400 via-purple-500 to-purple-600";
-  };
-
-  const NovelCard = ({ novel, rank }: { novel: Novel; rank: number }) => {
-    const styling = getNovelStyling(novel, "normal");
-    const badgeConfig = getNovelBadgeConfig(novel);
-    const isTopThree = rank <= 3;
-
-    return (
-      <Link href={`/novels/${novel.slug}`}>
-        <Card
-          className={cn(
-            "group relative cursor-pointer overflow-hidden transition-all duration-200 hover:scale-[1.01] hover:shadow-xl",
-            // Add featured/trending border styling
-            novel.is_featured &&
-              "border-2 border-amber-500/30 shadow-lg shadow-amber-500/10",
-            novel.is_trending &&
-              !novel.is_featured &&
-              "border-2 border-blue-500/30 shadow-lg shadow-blue-500/10",
-            // Add container gradient background
-            styling.containerClass,
-            // Special styling for top 3
-            isTopThree && "shadow-lg",
-          )}
-        >
-          <CardContent className="p-4">
-            <div className="flex gap-4">
-              {/* Rank Badge */}
-              <div className="flex flex-shrink-0 flex-col items-center gap-2">
-                <div
-                  className={cn(
-                    "relative flex h-10 w-10 items-center justify-center rounded-full text-sm font-bold text-white shadow-lg",
-                    `bg-gradient-to-br ${getRankColor(rank)}`,
-                    isTopThree && "shadow-xl ring-2 ring-white/50",
-                  )}
-                >
-                  {getRankIcon(rank) || rank}
-                </div>
-
-                {/* Cover Image */}
-                <div className="relative">
-                  <div className="h-[120px] w-[80px] flex-shrink-0 overflow-hidden rounded">
-                    {novel.cover_image ? (
-                      <Image
-                        src={novel.cover_image}
-                        alt={novel.title}
-                        width={80}
-                        height={120}
-                        className={cn(
-                          "bg-muted rounded object-cover transition-transform duration-300 group-hover:scale-105",
-                          styling.coverClass,
-                          isTopThree && "shadow-md",
-                        )}
-                      />
-                    ) : (
-                      <div
-                        className={cn(
-                          "bg-muted flex h-30 w-20 items-center justify-center rounded",
-                          styling.coverClass,
-                        )}
-                      >
-                        <BookOpen className="text-muted-foreground h-8 w-8" />
-                      </div>
-                    )}
-                  </div>
-
-                  {/* Corner badge icon */}
-                  {styling.showCornerIcon && (
-                    <div
-                      className={cn(
-                        "absolute -top-1 -right-1 rounded-full p-0.5 shadow-lg",
-                        styling.cornerIconClass,
-                      )}
-                    >
-                      {novel.is_featured ? (
-                        <Crown className="h-3 w-3 text-white" />
-                      ) : (
-                        <TrendingUp className="h-3 w-3 text-white" />
-                      )}
-                    </div>
-                  )}
-                </div>
-              </div>
-
-              {/* Novel Info */}
-              <div className="min-w-0 flex-1 space-y-2">
-                <div>
-                  <div className="flex flex-wrap items-center gap-2">
-                    <h3
-                      className={cn(
-                        "group-hover:text-primary line-clamp-2 text-lg font-semibold transition-colors",
-                        styling.titleClass,
-                      )}
-                    >
-                      {novel.title}
-                    </h3>
-                    {badgeConfig.show && (
-                      <NovelBadge
-                        novel={novel}
-                        positioned={false}
-                        className={styling.badge.className}
-                      />
-                    )}
-                  </div>
-                  <p className="text-muted-foreground">by {novel.author}</p>
-                </div>
-
-                <p className="text-muted-foreground line-clamp-2 text-sm">
-                  {truncateDescription(novel.description, 150)}
-                </p>
-
-                {/* Genres */}
-                <div className="flex flex-wrap gap-1">
-                  {novel.genres.slice(0, 3).map((genre) => (
-                    <Badge
-                      key={genre.id}
-                      variant="secondary"
-                      className="text-xs"
-                    >
-                      {genre.name}
-                    </Badge>
-                  ))}
-                  {novel.genres.length > 3 && (
-                    <Badge variant="secondary" className="text-xs">
-                      +{novel.genres.length - 3}
-                    </Badge>
-                  )}
-                </div>
-
-                {/* Stats */}
-                <div className="flex flex-wrap items-center gap-4 text-sm">
-                  <div className="flex items-center gap-1 font-medium text-yellow-600">
-                    {novel.rating !== null && novel.rating !== undefined && (
-                      <NovelRating novel={novel} />
-                    )}
-                  </div>
-                  <Badge
-                    variant={getStatusColor(novel.status)}
-                    className="text-xs"
-                  >
-                    {novel.status.charAt(0).toUpperCase() +
-                      novel.status.slice(1)}
-                  </Badge>
-                  <div className="text-muted-foreground flex items-center gap-1">
-                    <BookOpen className="h-4 w-4" />
-                    <span>{novel.total_chapters || 0} ch</span>
-                  </div>
-                  <div className="text-muted-foreground flex items-center gap-1">
-                    <Eye className="h-4 w-4" />
-                    <span>{formatViewCount(novel.views)}</span>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-      </Link>
-    );
-  };
 
   const LoadingSkeleton = ({ rank }: { rank: number }) => (
     <Card>
@@ -267,15 +72,15 @@ export default function TopRatedPage() {
         <div className="grid grid-cols-1 gap-4 md:grid-cols-3">
           {/* 2nd Place */}
           <div className="order-2 md:order-1">
-            <NovelCard novel={novels[1]} rank={2} />
+            <NovelCard novel={novels[1]} size="ranked" rank={2} />
           </div>
           {/* 1st Place */}
           <div className="order-1 md:order-2 md:-mt-4 md:scale-105">
-            <NovelCard novel={novels[0]} rank={1} />
+            <NovelCard novel={novels[0]} size="ranked" rank={1} />
           </div>
           {/* 3rd Place */}
           <div className="order-3 md:order-3">
-            <NovelCard novel={novels[2]} rank={3} />
+            <NovelCard novel={novels[2]} size="ranked" rank={3} />
           </div>
         </div>
       </div>
@@ -316,7 +121,12 @@ export default function TopRatedPage() {
         {novels.length > 3 && (
           <div className="space-y-4">
             {novels.slice(3).map((novel, index) => (
-              <NovelCard key={novel.id} novel={novel} rank={index + 4} />
+              <NovelCard
+                key={novel.id}
+                novel={novel}
+                size="ranked"
+                rank={index + 4}
+              />
             ))}
           </div>
         )}
