+++ src/components/novels/novel-card.tsx
@@ -21,19 +21,37 @@ import { NovelBadge } from "./ui/novel-badge";
 
 interface NovelCardProps {
   novel: Novel;
-  size?: "default" | "compact" | "featured" | "horizontal";
+  size?: "default" | "compact" | "featured" | "horizontal" | "ranked";
   className?: string;
+  rank?: number;
 }
 
 export function NovelCard({
   novel,
   size = "default",
   className,
+  rank,
 }: NovelCardProps) {
   const isCompact = size === "compact";
   const isFeatured = size === "featured";
   const isHorizontal = size === "horizontal";
+  const isRanked = size === "ranked";
   const styling = getNovelStyling(novel, "normal");
+
+  const getRankIcon = (rank: number) => {
+    if (rank === 1) return <div className="text-xl">1</div>;
+    if (rank === 2) return <div className="text-xl">2</div>;
+    if (rank === 3) return <div className="text-xl">3</div>;
+    return null;
+  };
+
+  const getRankColor = (rank: number) => {
+    if (rank === 1) return "from-yellow-400 via-yellow-500 to-yellow-600";
+    if (rank === 2) return "from-slate-300 via-slate-400 to-slate-500";
+    if (rank === 3) return "from-orange-400 via-orange-500 to-orange-600";
+    if (rank <= 10) return "from-blue-400 via-blue-500 to-blue-600";
+    return "from-purple-400 via-purple-500 to-purple-600";
+  };
 
   // Horizontal layout (cover left, content right)
   if (isHorizontal) {
@@ -147,6 +165,161 @@ export function NovelCard({
               </div>
             </div>
           </div>
+        </Card>
+      </Link>
+    );
+  }
+
+  // Ranked layout (with rank badge)
+  if (isRanked && rank !== undefined) {
+    const isTopThree = rank <= 3;
+
+    return (
+      <Link
+        href={`/novels/${novel.slug}`}
+        className="focus-visible:ring-ring block rounded-lg focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2"
+      >
+        <Card
+          className={cn(
+            "group relative overflow-hidden transition-all duration-200 hover:scale-[1.01] hover:shadow-xl",
+            // Add featured/trending border styling
+            novel.is_featured &&
+              "border-2 border-amber-500/30 shadow-lg shadow-amber-500/10",
+            novel.is_trending &&
+              !novel.is_featured &&
+              "border-2 border-blue-500/30 shadow-lg shadow-blue-500/10",
+            // Add container gradient background
+            styling.containerClass,
+            // Special styling for top 3
+            isTopThree && "shadow-lg",
+            className,
+          )}
+        >
+          <CardContent className="p-4">
+            <div className="flex gap-4">
+              {/* Rank Badge */}
+              <div className="flex flex-shrink-0 flex-col items-center gap-2">
+                <div
+                  className={cn(
+                    "relative flex h-10 w-10 items-center justify-center rounded-full text-sm font-bold text-white shadow-lg",
+                    `bg-gradient-to-br ${getRankColor(rank)}`,
+                    isTopThree && "shadow-xl ring-2 ring-white/50",
+                  )}
+                >
+                  {getRankIcon(rank) || rank}
+                </div>
+
+                {/* Cover Image */}
+                <div className="relative">
+                  <div className="relative h-[120px] w-[80px] flex-shrink-0 overflow-hidden rounded">
+                    {novel.cover_image ? (
+                      <Image
+                        src={novel.cover_image}
+                        alt={novel.title}
+                        fill
+                        className={cn(
+                          "object-cover transition-transform duration-300 group-hover:scale-105",
+                          styling.coverClass,
+                          isTopThree && "shadow-md",
+                        )}
+                        sizes="80px"
+                      />
+                    ) : (
+                      <div className="from-muted to-muted/50 flex h-full w-full items-center justify-center bg-gradient-to-br">
+                        <BookOpen className="text-muted-foreground h-8 w-8" />
+                      </div>
+                    )}
+                  </div>
+
+                  {/* Corner badge icon */}
+                  {styling.showCornerIcon && (
+                    <div
+                      className={cn(
+                        "absolute -top-1 -right-1 rounded-full p-0.5 shadow-lg",
+                        styling.cornerIconClass,
+                      )}
+                    >
+                      {novel.is_featured ? (
+                        <Crown className="h-3 w-3 text-white" />
+                      ) : (
+                        <TrendingUp className="h-3 w-3 text-white" />
+                      )}
+                    </div>
+                  )}
+                </div>
+              </div>
+
+              {/* Novel Info */}
+              <div className="min-w-0 flex-1 space-y-2">
+                <div>
+                  <div className="flex flex-wrap items-center gap-2">
+                    <h3
+                      className={cn(
+                        "group-hover:text-primary line-clamp-2 text-lg font-semibold transition-colors",
+                        styling.titleClass,
+                      )}
+                    >
+                      {novel.title}
+                    </h3>
+                    <NovelBadge
+                      novel={novel}
+                      positioned={false}
+                      className={styling.badge.className}
+                    />
+                  </div>
+                  <p className="text-muted-foreground">
+                    by {novel.author ?? "Anonymous"}
+                  </p>
+                </div>
+
+                <p className="text-muted-foreground line-clamp-2 text-sm">
+                  {truncateDescription(novel.description, 150)}
+                </p>
+
+                {/* Genres */}
+                <div className="flex flex-wrap gap-1">
+                  {novel.genres.slice(0, 3).map((genre) => (
+                    <Badge
+                      key={genre.id}
+                      variant="secondary"
+                      className="text-xs"
+                    >
+                      {genre.name}
+                    </Badge>
+                  ))}
+                  {novel.genres.length > 3 && (
+                    <Badge variant="secondary" className="text-xs">
+                      +{novel.genres.length - 3}
+                    </Badge>
+                  )}
+                </div>
+
+                {/* Stats */}
+                <div className="flex flex-wrap items-center gap-4 text-sm">
+                  <div className="flex items-center gap-1 font-medium text-yellow-600">
+                    {novel.rating !== null && novel.rating !== undefined && (
+                      <NovelRating novel={novel} />
+                    )}
+                  </div>
+                  <Badge
+                    variant={getStatusColor(novel.status)}
+                    className="text-xs"
+                  >
+                    {novel.status.charAt(0).toUpperCase() +
+                      novel.status.slice(1)}
+                  </Badge>
+                  <div className="text-muted-foreground flex items-center gap-1">
+                    <BookOpen className="h-4 w-4" />
+                    <span>{novel.total_chapters || 0} ch</span>
+                  </div>
+                  <div className="text-muted-foreground flex items-center gap-1">
+                    <Eye className="h-4 w-4" />
+                    <span>{formatViewCount(novel.views)}</span>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </CardContent>
         </Card>
       </Link>
     );
